// src/shared/types/db.ts
export namespace DB {
// ...
	export type ChuniStaticMusic = {
		id: number
		version: number
		songId: number | null
		chartId: number | null
		title: string | null
		artist: string | null
		level: number | null
		genre: string | null
		jacketPath: string | null
		worldsEndTag: string | null
		opt: bigint | null
	}
// ...
}